ak-cli: separate api crate for compile times
