+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "3"
 members = [
     "ak-agent",
     "ak-agent-desktop/src-tauri",
+    "ak-api-cli",
     "ak-api-cli-gen",
     "ak-browser-support",
     "ak-cli",
@@ -42,9 +43,11 @@ split-debuginfo = "packed"
 strip = "debuginfo"
 
 [workspace.dependencies]
+ak-api-cli = { path = "ak-api-cli" }
 ak-api-cli-gen = { path = "ak-api-cli-gen" }
-authentik-client = { git = "https://github.com/goauthentik/client-rust", version = "2026.11.0-rc1" , default-features = false, features = ["rustls"] }
+authentik-client = { git = "https://github.com/goauthentik/client-rust", rev="24c227964b05ff1c3cc9487ccdd40751ac89f6b9", version = "2026.11.0-rc1" , default-features = false, features = ["rustls"] }
 chrono = { version = "0.4.45", features = ["serde"] }
+clap = { version = "4.6.1", features = ["derive"] }
 color-eyre = "= 0.6.5"
 eyre = "= 0.6.14"
 hyper-util = "0.1.20"
