+++ ak-api-cli/Makefile
@@ -1,0 +1,5 @@
+include ../common.mk
+
+.PHONY: test
+test:
+	$(call cargo_test,ak-api-cli)
