+++ ak-cli/Cargo.toml
@@ -14,10 +14,11 @@ path = "src/main.rs"
 [dependencies]
 ak-platform = { path = "../ak-platform" }
 ak-meta = { path = "../ak-meta" }
+ak-api-cli = { workspace = true }
 aws-sdk-sts = { version = "1.106.0", features = ["behavior-version-latest"] }
 aws-types = "1.3.16"
 chrono = { workspace = true }
-clap = { version = "4.6.1", features = ["derive"] }
+clap = { workspace = true }
 clap_complete = "4"
 tracing = { workspace = true }
 oauth-device-flows = { workspace = true }
@@ -36,8 +37,3 @@ uuid.workspace = true
 reqwest.workspace = true
 sentry.workspace = true
 rmcp = { version = "3.0.1", features = ["transport-io"] }
-
-[build-dependencies]
-ak-api-cli-gen.workspace = true
-eyre.workspace = true
-cargo_metadata = "0.23"
